Tidy the movie scanner after its package extraction
The cutover moved the scanner cleanly but left the two scanners out of
sync with each other, and left some habits that only made sense while the
code lived in package main.

Shared primitives:
- Promote the music scanner's copy-on-write scanCache[K,V] to
  scanner.ScanCache and use it for the movie scanner's genre and artist
  memos. clone() cloned both maps per movie, and artistIDs grows across
  the whole scan, so the cost was O(movies x distinct artists).
- Drop the non-context WalkMediaLibrary; both scanners use the Context
  form and only its own tests called it.

TMDB path:
- Extract lookupTmdbMovie with early returns, replacing a five-deep
  block, and log search and detail failures. They were discarded
  silently, so a TMDB outage degraded a movie permanently -- the scan
  index then marks the file unchanged -- with no signal at all. A
  cancelled scan stays quiet.
- Probe before searching TMDB: ffprobe is required and TMDB search
  results are never cached.
- Rank once per file. summarizeTmdbCandidates now takes the ranked
  slice instead of re-ranking, the target title is normalized once per
  pass rather than per candidate, and the two title Replacers are built
  once rather than ~120 times per scanned file.

Dependencies:
- New defaults ScanContext, Wait, ScannerDBMu and both callbacks, so the
  scattered nil checks go away and a partially wired Scanner no longer
  panics on a nil mutex. Pair wait.Add/Done inside Start so runMovieScan
  stays callable on its own.

TMDB persistence:
- Collapse the five exported Process* shims into ApplyTmdbMetadata.
  IdentifyMovie re-implemented persistResolvedMovieTx's exact sequence;
  both now share one definition of what a TMDB match owns.

ffprobe:
- GetMetadata and GetAudioMetadata take a context and cap each probe at
  60s on top of it, so shutdown cancels an in-flight probe instead of
  waiting it out. Cancellation and timeout are reported distinctly.

Also split the 1754-line scanner_test.go to mirror the six source files
and drop two section banners left headless by the original split.
