+++ server/cmd/api/music_scanner.go
@@ -9,7 +9,6 @@ import (
 	"igloo/cmd/internal/helpers"
 	"igloo/cmd/internal/scanner"
 	spotifyapi "igloo/cmd/internal/spotify"
-	"maps"
 	"path/filepath"
 	"strconv"
 	"strings"
@@ -174,56 +173,6 @@ func (app *Application) loadMusicScanIndex(ctx context.Context) (map[string]int6
 // Scan context
 // ---------------------------------------------------------------------------
 
-// scanCache is a two-level map: transaction-local entries over an optional
-// read-only base layer. The scan-lifetime context uses just the local layer;
-// the per-track overlay created by clone() starts empty and reads through to
-// the scan layer. Discarding the overlay after a rolled-back transaction
-// discards its entries, which used to require cloning every map per track --
-// O(tracks x library) -- and now costs only the new entries.
-type scanCache[K comparable, V any] struct {
-	local map[K]V
-	base  map[K]V // nil on the scan-lifetime context
-}
-
-func newScanCache[K comparable, V any]() scanCache[K, V] {
-	return scanCache[K, V]{local: make(map[K]V)}
-}
-
-// overlay returns an empty transaction-local layer over this cache's entries.
-// Only meaningful on the scan-lifetime cache (base == nil); overlays are never
-// stacked.
-func (c scanCache[K, V]) overlay() scanCache[K, V] {
-	return scanCache[K, V]{local: make(map[K]V), base: c.local}
-}
-
-func (c scanCache[K, V]) get(k K) (V, bool) {
-	if v, ok := c.local[k]; ok {
-		return v, true
-	}
-	if c.base != nil {
-		if v, ok := c.base[k]; ok {
-			return v, true
-		}
-	}
-	var zero V
-	return zero, false
-}
-
-func (c scanCache[K, V]) has(k K) bool {
-	_, ok := c.get(k)
-	return ok
-}
-
-func (c scanCache[K, V]) set(k K, v V) {
-	c.local[k] = v
-}
-
-// mergeFrom publishes another cache's local entries into this cache's local
-// layer, after the transaction that wrote them committed.
-func (c scanCache[K, V]) mergeFrom(other scanCache[K, V]) {
-	maps.Copy(c.local, other.local)
-}
-
 type musicScanContext struct {
 	// trackIndex maps cleaned file path -> file size for every track already in
 	// the DB. It is read to skip unchanged files and is only written after a
@@ -234,24 +183,24 @@ type musicScanContext struct {
 	// scan. They are written inside the per-track transaction, so the clone
 	// overlay isolates them until commit to avoid caching ids from a rolled-back
 	// transaction.
-	musicianIDs scanCache[string, int64]
-	albumIDs    scanCache[string, int64]
-	genreIDs    scanCache[string, int64]
+	musicianIDs scanner.ScanCache[string, int64]
+	albumIDs    scanner.ScanCache[string, int64]
+	genreIDs    scanner.ScanCache[string, int64]
 	// musicianAlbums, musicianGenres, albumGenres and trackGenres remember which
 	// join rows this scan already wrote, keyed by musicIDPairKey. Like the id
 	// caches they are transaction-scoped until commit.
-	musicianAlbums scanCache[string, struct{}]
-	musicianGenres scanCache[string, struct{}]
-	albumGenres    scanCache[string, struct{}]
-	trackGenres    scanCache[string, struct{}]
+	musicianAlbums scanner.ScanCache[string, struct{}]
+	musicianGenres scanner.ScanCache[string, struct{}]
+	albumGenres    scanner.ScanCache[string, struct{}]
+	trackGenres    scanner.ScanCache[string, struct{}]
 	// spotifyArtistMisses and spotifyAlbumMisses cache unmatched/failed Spotify
 	// lookups so a scan queries Spotify at most once per artist or album.
-	spotifyArtistMisses scanCache[string, resolvedSpotifyMatch]
-	spotifyAlbumMisses  scanCache[string, resolvedSpotifyMatch]
+	spotifyArtistMisses scanner.ScanCache[string, resolvedSpotifyMatch]
+	spotifyAlbumMisses  scanner.ScanCache[string, resolvedSpotifyMatch]
 	// spotifyMusicianGenresHandled and spotifyAlbumGenresHandled record entities
 	// whose Spotify genres were fully written, so later tracks skip the work.
-	spotifyMusicianGenresHandled scanCache[int64, struct{}]
-	spotifyAlbumGenresHandled    scanCache[int64, struct{}]
+	spotifyMusicianGenresHandled scanner.ScanCache[int64, struct{}]
+	spotifyAlbumGenresHandled    scanner.ScanCache[int64, struct{}]
 }
 
 func newMusicScanContext(trackIndex map[string]int64) *musicScanContext {
@@ -263,49 +212,49 @@ func newMusicScanContext(trackIndex map[string]int64) *musicScanContext {
 	// and the caller discards its reference, so no defensive copy is needed.
 	return &musicScanContext{
 		trackIndex:                   trackIndex,
-		musicianIDs:                  newScanCache[string, int64](),
-		albumIDs:                     newScanCache[string, int64](),
-		genreIDs:                     newScanCache[string, int64](),
-		musicianAlbums:               newScanCache[string, struct{}](),
-		musicianGenres:               newScanCache[string, struct{}](),
-		albumGenres:                  newScanCache[string, struct{}](),
-		trackGenres:                  newScanCache[string, struct{}](),
-		spotifyArtistMisses:          newScanCache[string, resolvedSpotifyMatch](),
-		spotifyAlbumMisses:           newScanCache[string, resolvedSpotifyMatch](),
-		spotifyMusicianGenresHandled: newScanCache[int64, struct{}](),
-		spotifyAlbumGenresHandled:    newScanCache[int64, struct{}](),
+		musicianIDs:                  scanner.NewScanCache[string, int64](),
+		albumIDs:                     scanner.NewScanCache[string, int64](),
+		genreIDs:                     scanner.NewScanCache[string, int64](),
+		musicianAlbums:               scanner.NewScanCache[string, struct{}](),
+		musicianGenres:               scanner.NewScanCache[string, struct{}](),
+		albumGenres:                  scanner.NewScanCache[string, struct{}](),
+		trackGenres:                  scanner.NewScanCache[string, struct{}](),
+		spotifyArtistMisses:          scanner.NewScanCache[string, resolvedSpotifyMatch](),
+		spotifyAlbumMisses:           scanner.NewScanCache[string, resolvedSpotifyMatch](),
+		spotifyMusicianGenresHandled: scanner.NewScanCache[int64, struct{}](),
+		spotifyAlbumGenresHandled:    scanner.NewScanCache[int64, struct{}](),
 	}
 }
 
 func (scan *musicScanContext) clone() *musicScanContext {
 	return &musicScanContext{
 		trackIndex:                   scan.trackIndex, // shared; never written inside the transaction
-		musicianIDs:                  scan.musicianIDs.overlay(),
-		albumIDs:                     scan.albumIDs.overlay(),
-		genreIDs:                     scan.genreIDs.overlay(),
-		musicianAlbums:               scan.musicianAlbums.overlay(),
-		musicianGenres:               scan.musicianGenres.overlay(),
-		albumGenres:                  scan.albumGenres.overlay(),
-		trackGenres:                  scan.trackGenres.overlay(),
-		spotifyArtistMisses:          scan.spotifyArtistMisses.overlay(),
-		spotifyAlbumMisses:           scan.spotifyAlbumMisses.overlay(),
-		spotifyMusicianGenresHandled: scan.spotifyMusicianGenresHandled.overlay(),
-		spotifyAlbumGenresHandled:    scan.spotifyAlbumGenresHandled.overlay(),
+		musicianIDs:                  scan.musicianIDs.Overlay(),
+		albumIDs:                     scan.albumIDs.Overlay(),
+		genreIDs:                     scan.genreIDs.Overlay(),
+		musicianAlbums:               scan.musicianAlbums.Overlay(),
+		musicianGenres:               scan.musicianGenres.Overlay(),
+		albumGenres:                  scan.albumGenres.Overlay(),
+		trackGenres:                  scan.trackGenres.Overlay(),
+		spotifyArtistMisses:          scan.spotifyArtistMisses.Overlay(),
+		spotifyAlbumMisses:           scan.spotifyAlbumMisses.Overlay(),
+		spotifyMusicianGenresHandled: scan.spotifyMusicianGenresHandled.Overlay(),
+		spotifyAlbumGenresHandled:    scan.spotifyAlbumGenresHandled.Overlay(),
 	}
 }
 
 func (scan *musicScanContext) mergeFrom(other *musicScanContext) {
-	scan.musicianIDs.mergeFrom(other.musicianIDs)
-	scan.albumIDs.mergeFrom(other.albumIDs)
-	scan.genreIDs.mergeFrom(other.genreIDs)
-	scan.musicianAlbums.mergeFrom(other.musicianAlbums)
-	scan.musicianGenres.mergeFrom(other.musicianGenres)
-	scan.albumGenres.mergeFrom(other.albumGenres)
-	scan.trackGenres.mergeFrom(other.trackGenres)
-	scan.spotifyArtistMisses.mergeFrom(other.spotifyArtistMisses)
-	scan.spotifyAlbumMisses.mergeFrom(other.spotifyAlbumMisses)
-	scan.spotifyMusicianGenresHandled.mergeFrom(other.spotifyMusicianGenresHandled)
-	scan.spotifyAlbumGenresHandled.mergeFrom(other.spotifyAlbumGenresHandled)
+	scan.musicianIDs.MergeFrom(other.musicianIDs)
+	scan.albumIDs.MergeFrom(other.albumIDs)
+	scan.genreIDs.MergeFrom(other.genreIDs)
+	scan.musicianAlbums.MergeFrom(other.musicianAlbums)
+	scan.musicianGenres.MergeFrom(other.musicianGenres)
+	scan.albumGenres.MergeFrom(other.albumGenres)
+	scan.trackGenres.MergeFrom(other.trackGenres)
+	scan.spotifyArtistMisses.MergeFrom(other.spotifyArtistMisses)
+	scan.spotifyAlbumMisses.MergeFrom(other.spotifyAlbumMisses)
+	scan.spotifyMusicianGenresHandled.MergeFrom(other.spotifyMusicianGenresHandled)
+	scan.spotifyAlbumGenresHandled.MergeFrom(other.spotifyAlbumGenresHandled)
 }
 
 func (scan *musicScanContext) trackUnchanged(path string, size int64) bool {
@@ -357,7 +306,7 @@ type resolvedAlbum struct {
 }
 
 func (app *Application) resolveTrackFile(ctx context.Context, scan *musicScanContext, file scanner.ScanFile) (*resolvedTrack, error) {
-	info, err := app.Ffprobe.GetAudioMetadata(file.Path)
+	info, err := app.Ffprobe.GetAudioMetadata(ctx, file.Path)
 	if err != nil {
 		return nil, fmt.Errorf("ffprobe failed: %w", err)
 	}
@@ -515,7 +464,7 @@ func (app *Application) resolveTrackMusicians(ctx context.Context, scan *musicSc
 
 func (app *Application) resolveMusician(ctx context.Context, scan *musicScanContext, name, sortName string) (*resolvedMusician, error) {
 	cacheKey := scanner.NormalizedScanCacheKey(name, sortName)
-	if musicianID, ok := scan.musicianIDs.get(cacheKey); ok {
+	if musicianID, ok := scan.musicianIDs.Get(cacheKey); ok {
 		return &resolvedMusician{
 			name:          name,
 			sortName:      sortName,
@@ -542,7 +491,7 @@ func (app *Application) resolveMusician(ctx context.Context, scan *musicScanCont
 		if matchErr == nil {
 			if musicSpotifyMatchStatusIsFinal(persisted.Status) {
 				resolved.splitCompoundOnNoMatch = musicSpotifyMatchSplitsCompound(persisted.Status, persisted.Reason)
-				scan.musicianIDs.set(cacheKey, existing.ID)
+				scan.musicianIDs.Set(cacheKey, existing.ID)
 				return resolved, nil
 			}
 		} else if !errors.Is(matchErr, sql.ErrNoRows) {
@@ -551,7 +500,7 @@ func (app *Application) resolveMusician(ctx context.Context, scan *musicScanCont
 	}
 
 	spotifyKey := scanner.NormalizedScanCacheKey(name)
-	if cachedMiss, ok := scan.spotifyArtistMisses.get(spotifyKey); ok {
+	if cachedMiss, ok := scan.spotifyArtistMisses.Get(spotifyKey); ok {
 		resolved.spotifyMatch = &cachedMiss
 		resolved.splitCompoundOnNoMatch = musicSpotifyMatchSplitsCompound(cachedMiss.status, cachedMiss.reason)
 		return resolved, nil
@@ -559,7 +508,7 @@ func (app *Application) resolveMusician(ctx context.Context, scan *musicScanCont
 
 	if app.Spotify == nil {
 		if found {
-			scan.musicianIDs.set(cacheKey, existing.ID)
+			scan.musicianIDs.Set(cacheKey, existing.ID)
 		}
 		return resolved, nil
 	}
@@ -567,7 +516,7 @@ func (app *Application) resolveMusician(ctx context.Context, scan *musicScanCont
 	artist, err := app.Spotify.SearchArtistByName(ctx, name)
 	if err != nil {
 		match := resolvedSpotifyMatchFromError(err)
-		scan.spotifyArtistMisses.set(spotifyKey, match)
+		scan.spotifyArtistMisses.Set(spotifyKey, match)
 		resolved.spotifyMatch = &match
 		resolved.splitCompoundOnNoMatch = shouldSplitCompoundArtistCredits(err)
 		return resolved, nil
@@ -587,7 +536,7 @@ func (app *Application) resolveMusician(ctx context.Context, scan *musicScanCont
 
 func (app *Application) resolveAlbum(ctx context.Context, scan *musicScanContext, title, sortTitle, albumArtist string) (*resolvedAlbum, error) {
 	cacheKey := scanner.NormalizedScanCacheKey(title, albumArtist)
-	if albumID, ok := scan.albumIDs.get(cacheKey); ok {
+	if albumID, ok := scan.albumIDs.Get(cacheKey); ok {
 		return &resolvedAlbum{
 			title:         title,
 			sortTitle:     sortTitle,
@@ -618,7 +567,7 @@ func (app *Application) resolveAlbum(ctx context.Context, scan *musicScanContext
 		})
 		if matchErr == nil {
 			if musicSpotifyMatchStatusIsFinal(persisted.Status) {
-				scan.albumIDs.set(cacheKey, existing.ID)
+				scan.albumIDs.Set(cacheKey, existing.ID)
 				return resolved, nil
 			}
 		} else if !errors.Is(matchErr, sql.ErrNoRows) {
@@ -627,14 +576,14 @@ func (app *Application) resolveAlbum(ctx context.Context, scan *musicScanContext
 	}
 
 	spotifyKey := scanner.NormalizedScanCacheKey(title, albumArtist)
-	if cachedMiss, ok := scan.spotifyAlbumMisses.get(spotifyKey); ok {
+	if cachedMiss, ok := scan.spotifyAlbumMisses.Get(spotifyKey); ok {
 		resolved.spotifyMatch = &cachedMiss
 		return resolved, nil
 	}
 
 	if app.Spotify == nil {
 		if found {
-			scan.albumIDs.set(cacheKey, existing.ID)
+			scan.albumIDs.Set(cacheKey, existing.ID)
 		}
 		return resolved, nil
 	}
@@ -642,7 +591,7 @@ func (app *Application) resolveAlbum(ctx context.Context, scan *musicScanContext
 	albumDetails, err := app.Spotify.SearchAndGetAlbumDetails(ctx, title, albumArtist)
 	if err != nil {
 		match := resolvedSpotifyMatchFromError(err)
-		scan.spotifyAlbumMisses.set(spotifyKey, match)
+		scan.spotifyAlbumMisses.Set(spotifyKey, match)
 		resolved.spotifyMatch = &match
 		return resolved, nil
 	}
@@ -925,7 +874,7 @@ func (app *Application) persistResolvedTrackTx(ctx context.Context, qtx *databas
 
 func (app *Application) persistMusician(ctx context.Context, qtx *database.Queries, scan *musicScanContext, input resolvedMusician) (int64, error) {
 	cacheKey := scanner.NormalizedScanCacheKey(input.name, input.sortName)
-	if musicianID, ok := scan.musicianIDs.get(cacheKey); ok {
+	if musicianID, ok := scan.musicianIDs.Get(cacheKey); ok {
 		return musicianID, nil
 	}
 
@@ -1004,7 +953,7 @@ func (app *Application) persistMusician(ctx context.Context, qtx *database.Queri
 
 func (app *Application) persistAlbum(ctx context.Context, qtx *database.Queries, scan *musicScanContext, input resolvedAlbum) (int64, error) {
 	cacheKey := scanner.NormalizedScanCacheKey(input.title, input.albumArtist)
-	if albumID, ok := scan.albumIDs.get(cacheKey); ok {
+	if albumID, ok := scan.albumIDs.Get(cacheKey); ok {
 		return albumID, nil
 	}
 
@@ -1187,13 +1136,13 @@ func (app *Application) processSpotifyEntityGenres(
 	scan *musicScanContext,
 	entityID int64,
 	spotifyGenres []string,
-	handled scanCache[int64, struct{}],
+	handled scanner.ScanCache[int64, struct{}],
 	processor spotifyGenreProcessor,
 ) {
 	if len(spotifyGenres) == 0 {
 		return
 	}
-	if handled.has(entityID) {
+	if handled.Has(entityID) {
 		return
 	}
 
@@ -1219,13 +1168,13 @@ func (app *Application) processSpotifyEntityGenres(
 	}
 
 	if !hadError {
-		handled.set(entityID, struct{}{})
+		handled.Set(entityID, struct{}{})
 	}
 }
 
 func (app *Application) getOrCreateMusicGenreID(ctx context.Context, qtx *database.Queries, scan *musicScanContext, tag string) (int64, error) {
 	cacheKey := scanner.NormalizedScanCacheKey(tag, "music")
-	if genreID, ok := scan.genreIDs.get(cacheKey); ok {
+	if genreID, ok := scan.genreIDs.Get(cacheKey); ok {
 		return genreID, nil
 	}
 
@@ -1237,7 +1186,7 @@ func (app *Application) getOrCreateMusicGenreID(ctx context.Context, qtx *databa
 		return 0, err
 	}
 
-	scan.genreIDs.set(cacheKey, genre.ID)
+	scan.genreIDs.Set(cacheKey, genre.ID)
 	return genre.ID, nil
 }
 
@@ -1277,9 +1226,9 @@ func (app *Application) createTrackGenreIfNeeded(ctx context.Context, qtx *datab
 	})
 }
 
-func createCachedMusicRelationshipIfNeeded(cache scanCache[string, struct{}], leftID, rightID int64, create func() error) error {
+func createCachedMusicRelationshipIfNeeded(cache scanner.ScanCache[string, struct{}], leftID, rightID int64, create func() error) error {
 	cacheKey := musicIDPairKey(leftID, rightID)
-	if cache.has(cacheKey) {
+	if cache.Has(cacheKey) {
 		return nil
 	}
 
@@ -1288,7 +1237,7 @@ func createCachedMusicRelationshipIfNeeded(cache scanCache[string, struct{}], le
 		return err
 	}
 
-	cache.set(cacheKey, struct{}{})
+	cache.Set(cacheKey, struct{}{})
 	return nil
 }
 
@@ -1326,7 +1275,7 @@ func (app *Application) upsertMusicSpotifyMatchAndCacheID(
 	entityType string,
 	entityID int64,
 	match *resolvedSpotifyMatch,
-	cache scanCache[string, int64],
+	cache scanner.ScanCache[string, int64],
 	cacheKey string,
 ) error {
 	if match != nil {
@@ -1349,7 +1298,7 @@ func (app *Application) upsertMusicSpotifyMatchAndCacheID(
 		}
 	}
 
-	cache.set(cacheKey, entityID)
+	cache.Set(cacheKey, entityID)
 	return nil
 }
